hook up transform controls mode to UI
