+++ apps/client/app/play/playStore.ts
@@ -1,20 +1,13 @@
 import { WorldMetadata } from "@wired-protocol/types";
 import { create } from "zustand";
 
-import {
-  LeftPanelPage,
-  PlayMode,
-  RightPanelPage,
-  Tool,
-  WorldUriId,
-} from "./types";
+import { LeftPanelPage, PlayMode, RightPanelPage, WorldUriId } from "./types";
 
 export interface PlayStore {
   chatBoxFocused: boolean;
   leftPage: LeftPanelPage;
   mode: PlayMode;
   rightPage: RightPanelPage;
-  tool: Tool;
   metadata: WorldMetadata;
   uiAvatar: string;
   uiName: string;
@@ -27,7 +20,6 @@ export const usePlayStore = create<PlayStore>(() => ({
   metadata: { model: "" },
   mode: PlayMode.Play,
   rightPage: RightPanelPage.World,
-  tool: Tool.Translate,
   uiAvatar: "",
   uiName: "",
   worldId: { type: "uri", value: "" },
