+++ apps/client/src/play/ui/editor/Tools.tsx
@@ -1,14 +1,13 @@
 import * as ToggleGroup from "@radix-ui/react-toggle-group";
+import { Tool, useSceneStore } from "@unavi/engine";
 import { BiMove } from "react-icons/bi";
 import { CgArrowsExpandUpRight } from "react-icons/cg";
 import { MdSync } from "react-icons/md";
 
-import { usePlayStore } from "@/app/play/playStore";
-import { Tool } from "@/app/play/types";
 import Tooltip from "@/src/ui/Tooltip";
 
 export default function Tools() {
-  const tool = usePlayStore((state) => state.tool);
+  const tool = useSceneStore((state) => state.tool);
 
   return (
     <div className="fixed bottom-0 left-1/2 z-20 -translate-x-1/2 p-4">
@@ -17,13 +16,7 @@ export default function Tools() {
         value={tool}
         defaultValue={tool}
         onValueChange={(value) => {
-          switch (value) {
-            case Tool.Translate:
-            case Tool.Rotate:
-            case Tool.Scale:
-              usePlayStore.setState({ tool: value });
-              break;
-          }
+          useSceneStore.setState({ tool: value as Tool });
         }}
         className="flex items-center space-x-1 rounded-lg bg-neutral-900 p-1 text-white"
       >
@@ -49,7 +42,7 @@ interface ToolButtonProps {
 }
 
 function ToolButton({ tool, children }: ToolButtonProps) {
-  const selected = usePlayStore((state) => state.tool);
+  const selected = useSceneStore((state) => state.tool);
   const isSelected = selected === tool;
 
   return (
