+++ packages/engine/src/editor/index.ts
@@ -1,3 +1,4 @@
 export * from "./classes/SceneTree";
 export * from "./classes/TreeItem";
 export * from "./sceneStore";
+export * from "./types";
