+++ packages/engine/src/editor/types.ts
@@ -1,0 +1,5 @@
+export enum Tool {
+  Translate = "translate",
+  Rotate = "rotate",
+  Scale = "scale",
+}
